refactor(rust): split oversized modules outside pipeline and llm quirks
